Merge pull request #5267 from aneeshdk:Rs4ToMasterMerge

This change includes the following fixes
-	Added a new mechanism to mark reentrancy. When a helper calls into user code we will have to explicitly either set the implicit call flags or mark that call as reentrancy safe
-	Added checks in lowerer to confirm that we emit an implicit call bailout check after a helper call
-	Added failfasts, under a config flag, to validate the int range assumptions made in Globopt
-	Added check to make sure an opcode that can't have implicit call do not lower a helper call that can have implicit calls
-	Moved the reentrancy lock assert to Recycler right before dispose
-	Fixed an issue where during expression evaluation we were not restoring the updated block id in case of a parenthesis in destructed params
-	Along with ignoring null valuetype bit, also ignore undefined bit in the IsSimilar function as ToDefinite() strips out the undefined bit
-	Assert that we don't have a reentrancy lock when calling Dispose in a deterministic way.

+++ lib/Common/Memory/HeapBlockMap.cpp
@@ -1273,11 +1273,13 @@ void HeapBlockMap64::ForEachNodeInAddressRange(void * address, size_t pageCount,
 {
     uint lowerBitsAddress = ::Math::PointerCastToIntegralTruncate<uint>(address);
     uint nodePages = HeapBlockMap64::PagesPer4GB - lowerBitsAddress / AutoSystemInfo::PageSize;
+
     if (pageCount < nodePages)
     {
         nodePages = (uint)pageCount;
     }
 
+    // TODO: the loop is no longer needed as we are limiting the recycler object to be less than 2GB
     do
     {
         Node * node = FindNode(address);
@@ -1289,7 +1291,7 @@ void HeapBlockMap64::ForEachNodeInAddressRange(void * address, size_t pageCount,
         {
             break;
         }
-        address = (void *)((size_t)address + (nodePages * AutoSystemInfo::PageSize));
+        address = (void *)((size_t)address + ((size_t)nodePages * AutoSystemInfo::PageSize));
         nodePages = HeapBlockMap64::PagesPer4GB;
         if (pageCount < HeapBlockMap64::PagesPer4GB)
         {
